"Updated about section layout and fixed image alignment with AOS settings"
This commit message describes the changes made to the about section of the HTML file, where the layout of the about-content and about-img-wrapper has been updated and the alignment of the images has been fixed using AOS settings. The binary file changes in the image files suggest that their position or styling may have been adjusted in the new layout.