Update style.css to add overflow property to .section; remove ionicon CDN from index.html and add ionicon link instead; update theme-color in index.html
This commit updates the style.css file to add the 'overflow: hidden' property to the '.section' class. It also removes the ionicon CDN reference from the index.html file and adds a link to the ionicon library instead. Additionally, it updates the 'theme-color' meta tag in the index.html file to '#000000'.